Tweaking has occurred: - Adds a deprecated compatibility shim to support the old API - Do a minor version bump instead of a major one since the API is compatible - Splits up types into their own file - Mark _sendTopic as an internal function - Updates documentation - Fixes a bug where requests would be sent at the same time - Adds queue length property to topic connection to help with pooling implementations - Adds destroyed property to topic connection to help with pooling implementations
