+++ README.md
@@ -9,20 +9,43 @@ A communication layer between node.js and BYOND game servers.
 
 ### Example
 ```javascript
-const http2byond = require("./index.js");
-let connection = new http2byond({
-	timeout: 2000
-});
+const {createTopicConnection} = require("./index.js");
+let connection = createTopicConnection({
+  host: "localhost",
+  port: 6666
+})
 
-var form = {
-	ip: "localhost",
-	port: "6666",
-	topic: "?status"
-};
-
-connection.run(form).then((body) => {
-	console.log(body);
+connection.send("status").then((body) => {
+  console.log(body);
 }, (err) => {
-	console.error("ERR", err);
+  console.error("ERR", err);
 });
+
+async function anAsyncFunction() {
+  const result = connection.send("status");
+}
+```
+
+OR
+
+```javascript
+const {sendTopic} = require("./index.js");
+
+sendTopic({
+  host: "localhost",
+  port: 6666,
+  topic: "status"
+}).then((body) => {
+  console.log(body);
+}, (err) => {
+  console.error("ERR", err);
+});
+
+async function anAsyncFunction() {
+  const result = await sendTopic({
+    host: "localhost",
+    port: 6666,
+    topic: "status"
+  });
+}
 ```
