+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "http2byond",
   "description": "Communication layer between node.js and BYOND game servers.",
-  "version": "3.0.0",
+  "version": "2.1.0",
   "scripts": {
     "prepare": "npx tsc"
   },
